Removes the vlc://quit call from the playlist
This ensures that subsequent calls to the play utility always goes through, even if something else is already playing.

One issue is that subsequent calls to this program while an existing instance is running will cause the second program to simply cancel while the first continues with its new playlist. This can make process monitoring difficult.